Displayable width and height now in float, not double. Also cleaned up lots of unused imports and lack of generics in Displayable constructors, and more generics just about everywhere.

+++ ini/trakem2/display/Bucketable.java
@@ -9,6 +9,6 @@ public interface Bucketable {
 	public HashMap<Displayable, ArrayList<Bucket>> getBucketMap(Layer layer);
 	public void updateBucket(Displayable d, Layer layer);
 	public Project getProject();
-	public double getLayerWidth();
-	public double getLayerHeight();
+	public float getLayerWidth();
+	public float getLayerHeight();
 }
